Update .gitignore and enhance CRIU metrics and benchmarking tools
- Added Lambda build artifacts to .gitignore to prevent unnecessary files from being tracked.
- Enhanced CheckpointManager to support optional zstd-seekable compression for pages images during dump.
- Updated criu_metrics.py to parse additional fault-path bounded wait statistics from logs.
- Introduced new benchmarking tools: bench_s3_libcurl.c for S3 Range-GET throughput in C and bench_s3_range.py for a Python-based S3 Range-GET benchmark.
- Improved data initialization in dataproc_standalone.py to optimize memory usage.
- Removed redundant validation in matmul_workload.py for matrix size limits.

+++ workloads/dataproc_standalone.py
@@ -304,7 +304,9 @@ class DataSource:
         init_start = time.time()
         if HAS_NUMPY:
             np.random.seed(42)
-            self.data = np.random.randn(num_rows, num_cols).astype(np.float64)
+            # randn already returns float64; avoid a redundant .astype() that
+            # doubles peak memory during init (randn buffer + astype copy).
+            self.data = np.random.randn(num_rows, num_cols)
             # Add some structure: trends, correlations
             for c in range(min(5, num_cols)):
                 self.data[:, c] += np.linspace(0, 2, num_rows)  # trend
